Move encoding pkg to internal dir
