+++ internal/encoding/bytes.go
