+++ internal/encoding/bytes_test.go
