+++ internal/encoding/errors.go
