+++ internal/encoding/float.go
