+++ internal/encoding/float_test.go
