+++ internal/encoding/int_test.go
