+++ internal/encoding/null.go
