+++ internal/encoding/null_test.go
