+++ internal/encoding/string.go
