+++ internal/encoding/string_test.go
