+++ internal/encoding/type.go
