+++ internal/encoding/type_test.go
